Affinity: Move file under VMI
Currently, there is a dedicated package to calculate the required
affinity.

This was necessary when we had two objects of different kinds (pod, VMI).

Move the `affinity.go` file under the `vmi` package.
